Move validation into its group
